LPS-31444 Move logic from UpgradeJournalContent into UpgradeJournal to follow pattern in UpgradeScopes, clean up UpgradeScopes, and fix missing UpgradeProcess_6_2_0
