+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_6_2_0.java
@@ -22,9 +22,10 @@ import com.liferay.portal.upgrade.v6_2_0.UpgradeBlogsAggregator;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeCustomizablePortlets;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeDynamicDataListDisplay;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeDynamicDataMapping;
-import com.liferay.portal.upgrade.v6_2_0.UpgradeJournalContent;
+import com.liferay.portal.upgrade.v6_2_0.UpgradeJournal;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeMessageBoards;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeMessageBoardsAttachments;
+import com.liferay.portal.upgrade.v6_2_0.UpgradePortletPreferences;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeSchema;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeUser;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeWikiAttachments;
@@ -47,11 +48,12 @@ public class UpgradeProcess_6_2_0 extends UpgradeProcess {
 		upgrade(UpgradeBlogs.class);
 		upgrade(UpgradeBlogsAggregator.class);
 		upgrade(UpgradeCustomizablePortlets.class);
-		upgrade(UpgradeDynamicDataMapping.class);
 		upgrade(UpgradeDynamicDataListDisplay.class);
-		upgrade(UpgradeJournalContent.class);
+		upgrade(UpgradeDynamicDataMapping.class);
+		upgrade(UpgradeJournal.class);
 		upgrade(UpgradeMessageBoards.class);
 		upgrade(UpgradeMessageBoardsAttachments.class);
+		upgrade(UpgradePortletPreferences.class);
 		upgrade(UpgradeUser.class);
 		upgrade(UpgradeWikiAttachments.class);
 	}
