+++ portal-impl/src/com/liferay/portal/upgrade/v6_0_3/UpgradeScopes.java
@@ -28,12 +28,7 @@ public class UpgradeScopes extends BaseUpgradePortletPreferences {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-
-		// UUID
-
 		VerifyUUID.verifyModel("Layout", "plid");
-
-		// PortletPreferences
 
 		updatePortletPreferences();
 	}
