+++ portal-impl/src/com/liferay/portal/upgrade/v6_1_0/UpgradeScopes.java
@@ -26,11 +26,6 @@ import javax.portlet.PortletPreferences;
 public class UpgradeScopes extends BaseUpgradePortletPreferences {
 
 	@Override
-	protected void doUpgrade() throws Exception {
-		updatePortletPreferences();
-	}
-
-	@Override
 	protected String getUpdatePortletPreferencesWhereClause() {
 		return "preferences like '%lfrScopeLayoutUuid%'";
 	}
