--- portal-impl/src/com/liferay/portal/upgrade/v6_2_0/UpgradeJournalContent.java
@@ -1,44 +1,0 @@
-/**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.upgrade.v6_2_0;
-
-import com.liferay.portal.kernel.upgrade.RenameUpgradePortletPreferences;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * @author Marcellus Tavares
- */
-public class UpgradeJournalContent extends RenameUpgradePortletPreferences {
-
-	public UpgradeJournalContent() {
-		_preferenceNamesMap.put("templateId", "ddmTemplateKey");
-	}
-
-	@Override
-	protected String[] getPortletIds() {
-		return new String[] {"56_INSTANCE_%"};
-	}
-
-	@Override
-	protected Map<String, String> getPreferenceNamesMap() {
-		return _preferenceNamesMap;
-	}
-
-	private Map<String, String> _preferenceNamesMap =
-		new HashMap<String, String>();
-
-}
